Refactor: 커맨드 핸들러를 적용하여 StompHandler 리팩토링
